refactor: ReservationRepository, ReservationTimeRepository 구현
- 생성자 주입을 통한 의존성 관리
- 데이터 매핑 작업을 위한 메서드 추가
- AdminRepository 제거
