+++ src/main/java/roomescape/admin/repository/ReservationTimeRepository.java
@@ -1,26 +1,62 @@
 package roomescape.admin.repository;
 
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.stereotype.Repository;
 import roomescape.admin.dto.SaveReservationTimeRequest;
 import roomescape.admin.entity.ReservationTime;
 
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.util.List;
 
 @Repository
 public class ReservationTimeRepository {
-    public List<ReservationTime> readReservationTime() {
-        return null;
+
+    private final JdbcTemplate jdbcTemplate;
+
+    public ReservationTimeRepository(JdbcTemplate jdbcTemplate){
+        this.jdbcTemplate = jdbcTemplate;
     }
 
-    public Long saveReservationTime(SaveReservationTimeRequest saveReservationTimeRequest) {
-        return null;
+    public List<ReservationTime> readReservationTime() {
+        return this.jdbcTemplate.query(ReservationTimeQuery.READ_ALL, readReservationTimeRowMapper());
     }
 
     public ReservationTime readReservationTimeById(Long id) {
-        return null;
+        return this.jdbcTemplate.queryForObject(ReservationTimeQuery.READ_BY_ID, readReservationTimeRowMapper(),id);
+    }
+
+    private RowMapper<ReservationTime> readReservationTimeRowMapper() {
+        return (resultSet, rowNum) -> mapToReservationTime(resultSet);
+    }
+
+    private ReservationTime mapToReservationTime(ResultSet rs) throws SQLException {
+        long id = rs.getLong("time_id");
+        String startAt = rs.getString("time_start_at");
+
+        return ReservationTime.of(id, startAt);
+    }
+
+    public Long saveReservationTime(SaveReservationTimeRequest saveReservationTimeRequest) {
+        GeneratedKeyHolder keyHolder = new GeneratedKeyHolder();
+        this.jdbcTemplate.update(con -> saveReserveTimeStatement(con, saveReservationTimeRequest), keyHolder);
+        return keyHolder.getKey().longValue();
+    }
+
+    private PreparedStatement saveReserveTimeStatement(Connection con, SaveReservationTimeRequest saveReservationTimeRequest) throws SQLException {
+        String startAt = saveReservationTimeRequest.startAt();
+        PreparedStatement ps = con.prepareStatement(ReservationTimeQuery.SAVE, new String[]{"id"});
+        ps.setString(1, startAt);
+
+        return ps;
     }
 
     public void deleteReservationTime(Long id) {
-
+        this.jdbcTemplate.update(ReservationTimeQuery.DELETE, id);
     }
+
 }
